hexatodec correct

+++ function.c
@@ -27,8 +27,8 @@ int hexatodec(char *val){
     int result=0;
     int powcalc=1;
     int valSym = 0;
-    long unsigned int valsize = sizeof(val) / sizeof(char);
-    printf("\nTAILLE TABLEAU IMPORTANT : %lu\n", valsize);
+    int valsize = sizeof(val) / sizeof(char);
+    printf("\nTAILLE TABLEAU IMPORTANT : %d\n", valsize);
     for(int i=0; i<sizeof(val);i++){
         int base=1; 
         switch(val[i]){
